requirements.txt
asgiref==3.5.2
Django==4.1.2
django-crispy-forms==1.14.0
Pillow==9.2.0
psycopg2-binary==2.9.4
python-decouple==3.6
sqlparse==0.4.3
